ENH: Cython interface for special
This defines typed, scalar versions of many of the functions in
scipy.special and makes them available to cimport in Cython. It uses
fused types so that the functions have the same names and signatures
as their ufunc counterparts.

It piggybacks off of the existing code generation in
generate_ufuncs.py. In particular:

- The process for adding new ufuncs is unchanged
- When a new ufunc is added a Cython version will automatically be
  created (subject to some restrictions mentioned at the end), and the
  documentation and tests will be automatically updated.

This does not attempt to add Cython versions of:

- Ufuncs with multiple return arguments
- Ufuncs that have additional layers of python wrappers.

With more work these restrictions could be removed.

Closes gh-5964.